# Conflicts:
#	src/main/java/org/bimrocket/ihub/connector/Connector.java
#	src/main/java/org/bimrocket/ihub/service/ConnectorService.java
#	src/main/java/org/bimrocket/ihub/util/ConfigPropertyHandler.java

+++ src/main/java/org/bimrocket/ihub/service/ConnectorService.java
@@ -1,4 +1,4 @@
-/*
+/**
  * BIMROCKET
  *
  * Copyright (C) 2022, Ajuntament de Sant Feliu de Llobregat
@@ -34,15 +34,17 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.regex.Pattern;
+
 import org.bimrocket.ihub.connector.Connector;
-import org.bimrocket.ihub.exceptions.InvalidSetupException;
+import org.bimrocket.ihub.exceptions.InvalidNameException;
 import org.bimrocket.ihub.exceptions.NotFoundException;
+import org.bimrocket.ihub.repo.ConnectorSetupRepository;
 import org.bimrocket.ihub.repo.IdPairRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.event.ApplicationReadyEvent;
 import org.springframework.context.ApplicationListener;
 import org.springframework.stereotype.Service;
-import org.bimrocket.ihub.repo.ConnectorSetupRepository;
 
 /**
  *
@@ -50,10 +52,10 @@ import org.bimrocket.ihub.repo.ConnectorSetupRepository;
  */
 @Service
 public class ConnectorService
-    implements ApplicationListener<ApplicationReadyEvent>
+  implements ApplicationListener<ApplicationReadyEvent>
 {
-  private final Map<String, Connector> connectors = Collections
-      .synchronizedMap(new HashMap<>());
+  private final Map<String, Connector> connectors =
+    Collections.synchronizedMap(new HashMap<>());
 
   @Autowired
   IdPairRepository idPairRepository;
@@ -64,12 +66,19 @@ public class ConnectorService
   @Autowired
   ConnectorMapperService connectorMapperService;
 
+  private final Pattern connectorNamePattern =
+    Pattern.compile("^([a-zA-Z_$][a-zA-Z\\d_$]*)$");
+
   public Connector createConnector(String connectorName)
-      throws InvalidSetupException
+    throws InvalidNameException
   {
     if (connectors.containsKey(connectorName))
-      throw new InvalidSetupException(310, "Connector name {%s} already exists",
-          connectorName);
+      throw new InvalidNameException(310,
+        "Connector name {%s} already exists", connectorName);
+
+    if (!connectorNamePattern.matcher(connectorName).matches())
+      throw new InvalidNameException(312,
+        "Invalid connector name {%s}", connectorName);
 
     Connector connector = new Connector(this, connectorName);
 
@@ -85,26 +94,46 @@ public class ConnectorService
     {
       connector.stop();
 
-      if (deleteFromRepo)
-        connectorSetupRepository.deleteById(connectorName);
+      if (deleteFromRepo) connectorSetupRepository.deleteById(connectorName);
 
       return true;
     }
     return false;
   }
 
-  public Connector getConnector(String connectorName) throws NotFoundException
+  public boolean hasConnector(String connectorName)
+  {
+    return connectors.containsKey(connectorName);
+  }
+
+  public Connector getConnector(String connectorName)
+    throws NotFoundException
   {
     Connector connector = connectors.get(connectorName);
     if (connector == null)
-      throw new NotFoundException(150, "Connector {%s} not found",
-          connectorName);
+      throw new NotFoundException(150,
+        "Connector {%s} not found", connectorName);
     return connector;
   }
 
   public ArrayList<Connector> getConnectors()
   {
     return new ArrayList<>(connectors.values());
+  }
+
+  public ArrayList<Connector> getConnectorsByName(String name)
+  {
+    ArrayList<Connector> filteredConnectors = new ArrayList<>();
+    String lowerCaseName = name == null ? null : name.toLowerCase();
+    connectors.values().forEach(connector ->
+    {
+      if (name == null
+        || connector.getName().toLowerCase().contains(lowerCaseName))
+      {
+        filteredConnectors.add(connector);
+      }
+    });
+    return filteredConnectors;
   }
 
   public IdPairRepository getIdPairRepository()
@@ -130,8 +159,7 @@ public class ConnectorService
       {
         System.out.println("Restoring " + connSetup.getName());
         Connector connector = createConnector(connSetup.getName()).restore();
-        if (connector.isAutoStart())
-          connector.start();
+        if (connector.isAutoStart()) connector.start();
       }
       catch (Exception ex)
       {
@@ -141,7 +169,7 @@ public class ConnectorService
   }
 
   @Override
-  public void onApplicationEvent(ApplicationReadyEvent event)
+	public void onApplicationEvent(ApplicationReadyEvent event)
   {
     System.out.println("INIT");
     System.out.println("idPairRepository: " + idPairRepository);
